Move ADT's to adt directory.
