+++ lib/trees/binary_tree.dart
@@ -1,5 +1,5 @@
 // @dart=2.9
-import 'binary_tree_adt.dart';
+import 'adt/binary_tree_adt.dart';
 
 /// A binary node has at most two children, which are referred to as the
 ///  [left] child and the [right] child.
