+++ lib/trees/red_black_tree.dart
@@ -1,5 +1,5 @@
 // @dart=2.9
-import './binary_tree_adt.dart';
+import 'adt/binary_tree_adt.dart';
 import 'binary_search_tree.dart';
 import 'binary_tree.dart';
